Improve custom artifact script handling.
This change will make it so api server will fail to deploy if a custom
artifact script provided by the user does not exist. Furthermore, it
will also ensure that the default artifact script follows naming
conventions that are consistent with the rest of the namespace scoped
resources. Tests are addded to accomodate this change.

Signed-off-by: Humair Khan <humair88@hotmail.com>