Removed omitempty from bool fields.
Omitempty will omit "false" bool fields, this is problematic when we
want to set default fields for bool fields. The unintendend consequence
of setting both means a "false" field will be omitted, but then be
replaced by the "default" field if the default is "true". As such we
remove omitempty and set bool fields to optional.

Signed-off-by: Humair Khan <humair88@hotmail.com>