Parameterize DSPA default images.
This change introduces the ability to pass in default images via either
a configmap or env vars. Environment variables take precedence over
configmaps. Environment variables follow the same nested behavior as the
yaml config equivalent. For example "images.apiserver" in a
`config.yaml` passed in to the operator is the equivalent to passing in
an ENV Variable `IMAGES_APISERVER`.