Add tests for operator config file and images.
This change updates existing tests to consume DSPO config files as part
of it's testing. For test cases that use default settings, the config
files are used. For cases where cr specifies specific images, the tests
consume the config but verify that it is not used over those specified
in the CR.