removed `replicas: 1` from mariadb deployment
Setting `replicas: 1` on the mariadb deployment causes problems
in our sandbox environment where an idler scales down workloads
to save costs. The idler and the ReplicaSet fight, which causes
the mariadb pod to start and stop infinitely, which eventually
leads to InnoDB corruption. For now, solve by not setting
`replicas: 1` so that the idler wins.

Fixes: https://issues.redhat.com/browse/RHODS-7771