Make object store port/secure scheme optional.
For cloud storage like aws s3, the secure port is 443, but this does not
work with the minio module that is used by dsp api server, so instead in
such a case we make port optional, and do not require it. When an empty
port is provided, then it is simply not used by dsp api server.

Test cases are provided to reflect this change.

Docs have also been provided on how to utilize this external object
storage with an aws s3 specific example usecase.

Signed-off-by: Humair Khan <humair88@hotmail.com>