Fix artifact passing for recurring runs.
Due to a bug in upstream code, scheduled run pipelines are not getting
the k8s annotations on the tekton pods that contain the artifact
endpoint and bucket info. As such artifact passing works in regular
(non-scheduled) runs but not for recurring runs, failing at the
copy-artifacts container step.

Hardcoding these values within the artifact script serves as a
sufficient work around for this problem.

Signed-off-by: Humair Khan <humair88@hotmail.com>