Add release automation workflow.
This change adds the necessary gh workfows to automate releases for
minor releases.

Workflows are split into 3 components:
1. Prep
2. Tests
3. Release Creation

1. Prep
Prep workflow will validate compatibilty.yaml, and ensure the minor
branches (e.g. 1.3.x) are not already created. If branches already
exist, the Prep workflow will fail. If comptability.yaml does not
contain the minor version entry, the workflow will fail. If
compatibility.yaml contains the minor version entry, but comptaiblity.md
is not up to date, then workflow will create the necessary PR and exit
early, notifying the user to retry the Prep workflow once this PR is
merged and compatiblity docs are in sync.

If all pre-requisites are met, Prep workflow will create the necessary
minor branches in DSP and DSPO repositories, then continue to perform
all the necessary image builds for DSPO + DSP components. Once the
builds are done, the Prep workfow will generate the necessary follow up
PR to the minor release branch, updating the params.env with the SHAs of
the newly built images.

2. Tests
This is a placeholder workflow to add future tests required to perform a
stable release. These are ci tests that run on the PR created via Prep
workflow above. Once the CI tests pass, the PR will need to be manually
merged, the workflow does not auto merge PRs (nor does it auto push
commits to any branches).

3. Release Creation
Once the PR from (1) is merged. A trigger workflow is launched, this
workflow exists to trigger the "Create Release" workflow. The trigger
workflow will package required information that is pulled from the PR
body (in the form of a yaml) and will upload it to GH workflow cache,
for the "Create Release" workflow to utilize. Once the trigger workflow
completes, the "Create Release" worklfow will begin.

The "Create Release" workflow Validates the triggering PR from (1), this
step is to prevent unauthorized entities from triggering this workflow.
Validation confirms the PR info being passed into this workflow
originates from a PR authored by a user in the "approvers" list in DSPO
repo's OWNERS file. Once validated, "Create Release" workflow will
create releases in DSP/DSPO, providing links to these releases in the
workflow summary, as well as a PR comment to the PR created in (1).

Signed-off-by: Humair Khan <HumairAK@users.noreply.github.com>