Switch from OAuth Proxy to Kube RBAC Proxy
In ROSA clusters and some OpenShift configurations, there won't be an
internal OAuth server. Besides, DSP does not actually leverage OAuth
cookies and relies on the user to provide their Kubernetes token
in an authorization header.

This also switches the RBAC to a more secure form to take advantage of
the major version update of ODH/RHOAI and that kube-rbac-proxy requires
mapping HTTP methods to verbs (e.g. POST is create), unlike OAuth proxy.

Note this removes the KFP frontend from DSPO because OAuth Proxy was
required to generate a cookie with a Kubernetes token to pass to the
backend.

Relates:
https://issues.redhat.com/browse/RHOAIENG-34577

Signed-off-by: mprahl <mprahl@users.noreply.github.com>